tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/type.replace.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: type.replace</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>type.replace</h1>
<p class="desc">HTMLCanvasElement methods can be replaced, and the replacement methods used by canvases</p>

<p class="notes">Defined in "Web IDL" (draft)
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("HTMLCanvasElement methods can be replaced, and the replacement methods used by canvases");
_addTest(function(canvas, ctx) {

window.HTMLCanvasElement.prototype.getContext = function (name) { return 0; };
_assertSame(canvas.getContext('2d'), 0, "canvas.getContext('2d')", "0");


});
</script>
